+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
-## 1.0.0
+## v1.1.0
+
+`PerformanceIndex` is added.
+
+Units are added with some conversion methods:
+
+- Fahrenheit - Celsius
+- Meter - centimeter, kilometer, inch, mile
+- Meter/Second - kilometer/hour, mile/hour
+- Radian - degree
+- Watt - kilowatt, horsepower
+- Newtonmeter - footpound
+- Pound/SuqaredInch - bar
+- Second, Millisecond - `time.Duration`
+
+## v1.0.0
 
 Initial release